+++ src/components/Button.jsx
@@ -1,0 +1,4 @@
+const Button = ({role, name, fnc, para}) =>{
+    return <button className={role} onClick={() => fnc(para.id)}>{name}</button>
+};
+ export default Button;
