--- src/components/Buttons.jsx
@@ -1,7 +1,0 @@
-import React, { Component } from 'react';
-
-const AddButton = () =>{
-    return <button className='add-btn'>추가하기</button>
-};
-
-export default AddButton;
