--- src/components/ToDoBox.jsx
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-import App from '../App';
-
-const ToDoBox = (props) => {
-
-};
-
-
-export default ToDoBox
